requirements.txt
tensorboard==2.9
tensorboardX
jiwer
soundfile
ffmpeg<5
addict
torchaudio
librosa
datasets==2.8.0
evaluate
transformers
mecab-python3
unidic-lite
pykakasi
dill<0.3.5